pyproject.toml
[project]
name = "agent-starter-pack"
version = "0.4.4"
description = "CLI to bootstrap production-ready Google Cloud GenAI agent projects from templates."
authors = [
    { name = "Google LLC", email = "agent-starter-pack@google.com" },
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = ["google-cloud-core (>=2.4.3,<3.0.0)", "google-auth (>=2.40.2,<3.0.0)", "google-cloud-aiplatform[adk,agent-engines] (==1.95.1)", "python-dotenv>=0.21.0", "supabase>=1.0.0"]

[dependency-groups]
dev = [
    "black~=23.11.0",
    "crewai~=0.95.0",
    "flake8~=6.1.0",
    "google-cloud-logging~=3.11.4",
    "grpcio~=1.70.0",
    "isort~=5.12.0",
    "langchain-core~=0.3.34",
    "langchain-google-community[vertexaisearch]~=2.0.4",
    "langchain-google-vertexai~=2.0.13",
    "langgraph~=0.2.72",
    "myst-parser~=2.0.0",
    "opentelemetry-instrumentation-langchain~=0.38.5",
    "pytest~=7.4.3",
    "pytest-cov~=4.1.0",
    "pytest-mock~=3.12.0",
    "pytest-rerunfailures>=15.0",
    "pytest-xdist~=3.6.1",
    "rich~=13.7.0",
    "sphinx~=7.1.2",
    "sphinx-autoapi~=3.0.0",
    "sphinx-click~=5.1.0",
    "sphinx-rtd-theme~=2.0.0",
    "streamlit~=1.42.0",
    "streamlit-feedback~=0.1.4",
    "watchdog~=6.0.0",
]

[project.optional-dependencies]
jupyter = [
    "ipykernel>=6.29.5",
    "jupyter"
]

lint = [
    "ruff>=0.4.6",
    "mypy~=1.15.0",
    "codespell~=2.2.0",
    "types-pyyaml~=6.0.12.20240917",
    "types-requests~=2.32.0.20240914",
]

[tool.pytest.ini_options]
pythonpath = [".", "src/frontends/streamlit", "src", "agents/agentic_rag"]
testpaths = ["tests"]
addopts = "-s -v --ignore=tests/integration/test_template_linting.py --ignore=tests/integration/test_templated_patterns.py"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s - %(levelname)s - %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.ruff]
line-length = 88
target-version = "py310"
include = ["./agents/**/*.py", "src/cli/**/*.py", "tests/**/*.py", "src/frontends/streamlit/**/*.py", "src/frontends/streamlit_adk/**/*.py"]
exclude = ["./agents/agentic_rag/**/*.py"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901", "RUF001"]

[tool.ruff.lint.isort]
known-first-party = ["app", "frontend"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
exclude = '''(?x)(
    ^agents/agentic_rag/.*$  # exclude all files under agentic_rag directory
)'''

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,**/*.ipynb"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
agent-starter-pack = "src.cli.main:cli"

[tool.poetry]
name = "agent-starter-pack"
version = "0.4.4"
description = "CLI to bootstrap production-ready Google Cloud GenAI agent projects from templates."
authors = ["Google LLC <agent-starter-pack@google.com>"]
packages = [
    { include = "agents", from = "src" },
    { include = "cli", from = "src" },
    { include = "data_ingestion", from = "src" },
    { include = "deployment_targets", from = "src" },
    { include = "frontends", from = "src" },

    { include = "utils", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
click = "~=8.1.7"
cookiecutter = "~=2.5.0"
google-genai = ">=1.18.0"
protobuf = "==5.29.4"
httpx = ">=0.28.1,<1"
rich = "~=13.7.0"
pyyaml = "~=6.0.1"
backoff = ">=2.2.1"
supabase = ">=2.15.1,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
google-cloud-aiplatform = {version = "1.95.1"}
grpcio = ">=1.63.0,<1.64"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[tool.uv.workspace]
members = ["temp"]
